+++ hal/inc/core_hal.h
@@ -123,6 +123,16 @@ void HAL_Bootloader_Lock(bool lock);
 
 bool HAL_Core_System_Reset_FlagSet(RESET_TypeDef resetType);
 
+
+typedef struct runtime_info_t {
+    uint16_t size;              /* Size of this struct. */
+    uint16_t flags;             /* reserved, set to 0. */
+    uint32_t freeheap;          /* Amount of guaranteed heap memory available. */
+    uint32_t system_version;
+} runtime_info_t;
+
+uint32_t HAL_Core_Runtime_Info(runtime_info_t* info, void* reserved);
+
 extern void app_setup_and_loop();
 
 #ifdef __cplusplus