+++ hal/inc/hal_dynalib_core.h
@@ -55,6 +55,7 @@ DYNALIB_FN(hal_core,HAL_Notify_WDT)
 DYNALIB_FN(hal_core,HAL_Bootloader_Get_Flag)
 DYNALIB_FN(hal_core,HAL_Bootloader_Lock)
 DYNALIB_FN(hal_core,HAL_Core_System_Reset_FlagSet)
+DYNALIB_FN(hal_core,HAL_Core_Runtime_Info)
 
 DYNALIB_END(hal_core)